improved readme and installation script

+++ traze-client/install.sh
@@ -6,10 +6,14 @@ ENV_NAME=traze
 
 sudo apt-get update && sudo apt-get install -y build-essential cmake wget
 
-wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
-bash Miniconda3-latest-Linux-x86_64.sh
-echo 'export PATH="~/miniconda3/bin:$PATH"' >> ~/.bashrc && source ~/.bashrc
-rm -f Miniconda3-latest-Linux-x86_64.sh
+echo "Should a new version of Miniconda be installed to your system?"
+read -p "Press the key 'n' only if you already have it! [[Y]/n] " -n 1 -r && echo # move to a new line
+if [[ ! $REPLY = [Nn] ]]; then
+	wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
+	bash Miniconda3-latest-Linux-x86_64.sh
+	echo 'export PATH="~/miniconda3/bin:$PATH"' >> ~/.bashrc && source ~/.bashrc
+	rm -f Miniconda3-latest-Linux-x86_64.sh
+fi
 
 conda create -yn ${ENV_NAME} python=${PYTHON_VERSION}
 source activate ${ENV_NAME}
@@ -17,7 +21,7 @@ source activate ${ENV_NAME}
 conda install -y cython
 
 wget -qO- https://github.com/nest/nest-simulator/archive/v${NEST_VERSION}.tar.gz | tar xvz
-cd nest-simulator-${NEST_VERSION}
+mv nest-simulator-${NEST_VERSION} nest-simulator && cd nest-simulator
 
 mkdir build && cd build
 cmake -Dwith-python=3 -DCMAKE_INSTALL_PREFIX:PATH=$PWD ..
